@@ -21,8 +21,8 @@
     "@iconify/svelte": "^4.0.2",
     "@playwright/test": "^1.46.1",
     "@sveltejs/adapter-cloudflare-workers": "^2.5.6",
-    "@sveltejs/kit": "^2.5.24",
-    "@sveltejs/vite-plugin-svelte": "^3.1.2",
+    "@sveltejs/kit": "^2.5.27",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@types/eslint": "^8.56.11",
     "@typescript-eslint/eslint-plugin": "^7.18.0",
     "@typescript-eslint/parser": "^7.18.0",
@@ -30,17 +30,17 @@
     "drizzle-kit": "^0.22.7",
     "eslint": "^8.56.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-svelte": "^2.35.1",
+    "eslint-plugin-svelte": "^2.45.1",
     "mdsvex": "^0.11.0",
     "postcss": "^8.4.41",
     "prettier": "^3.1.1",
-    "prettier-plugin-svelte": "^3.1.2",
-    "svelte": "^4.2.7",
-    "svelte-check": "^3.8.5",
+    "prettier-plugin-svelte": "^3.2.6",
+    "svelte": "^5.0.0",
+    "svelte-check": "^4.0.0",
     "tailwindcss": "^3.4.10",
     "tslib": "^2.4.1",
     "typescript": "^5.5.4",
-    "vite": "^5.4.2",
+    "vite": "^5.4.4",
     "vitest": "^1.2.0"
   },
   "type": "module",
