@@ -3,9 +3,12 @@
     import { DateTime } from 'luxon';
     import { DateUtils } from '$lib/date';
 
-    export let value: string;
-    let v = [DateUtils.getDate(new Date(value)), DateUtils.getTime(new Date(value))];
-    $: v = [DateUtils.getDate(new Date(value)), DateUtils.getTime(new Date(value))];
+    interface Props {
+        value: string;
+    }
+
+    let { value = $bindable() }: Props = $props();
+    let v = $derived([DateUtils.getDate(new Date(value)), DateUtils.getTime(new Date(value))]);
 
     const onselect = () => {
         value = `${v[0]}T${v[1]}:00${DateTime.now().setZone(conf.timezone).toFormat("ZZ")}`;
@@ -16,7 +19,7 @@
 </script>
 
 <div class="w-full grid grid-cols-3 gap-2">
-    <select bind:value={v[0]} on:change={onselect} class="col-span-1">
+    <select bind:value={v[0]} onchange={onselect} class="col-span-1">
         {#each DateUtils.eachDays(DateUtils.defaultDate(conf.start, conf.end), conf.end) as d}
             <option value={d?.toFormat("yyyy-MM-dd")}>
                 {d?.toFormat("M/d")}
@@ -24,5 +27,5 @@
         {/each}
     </select>
 
-    <input type="time" step={60} on:change={oninput} bind:value={v[1]} class="col-span-2">
+    <input type="time" step={60} onchange={oninput} bind:value={v[1]} class="col-span-2">
 </div>
